fix: local debug for spfx
